♻️ refactor: refactor message group
