+++ packages/database/src/schemas/_helpers.ts
@@ -1,6 +1,8 @@
-import { timestamp } from 'drizzle-orm/pg-core';
+import { timestamp, varchar } from 'drizzle-orm/pg-core';
 
 export const timestamptz = (name: string) => timestamp(name, { withTimezone: true });
+
+export const varchar255 = (name: string) => varchar(name, { length: 255 });
 
 export const createdAt = () => timestamptz('created_at').notNull().defaultNow();
 export const updatedAt = () =>
