+++ packages/database/src/schemas/relations.ts
@@ -9,8 +9,7 @@ import { chatGroups, chatGroupsAgents } from './chatGroup';
 import { documentChunks, documents } from './document';
 import { files, knowledgeBases } from './file';
 import { generationBatches, generationTopics, generations } from './generation';
-import { messages, messagesFiles } from './message';
-import { messageGroups } from './messageGroup';
+import { messageGroups, messages, messagesFiles } from './message';
 import { chunks, unstructuredChunks } from './rag';
 import { sessionGroups, sessions } from './session';
 import { threads, topicDocuments, topics } from './topic';
